+++ R/calc_movement_vectors.R
@@ -56,7 +56,7 @@ calc_movement_vectors <- function(bf, start, direction = "forward") {
   }
   end <- lookup_timestep(end, bf)
 
-  trans_name <- lookup_transitions(bf, start, end, direction)
+  trans_name <- lookup_transitions(bf, start = start, end = end, direction = direction)
   stopifnot(length(trans_name) == 1)
 
   trans <- get_transition(bf, trans_name)